requirements.txt
attrs==17.2.0
cattrs==0.4.0
certifi==2017.4.17
chardet==3.0.4
flake8==3.3.0
idna==2.5
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
Pillow==4.2.0
pycodestyle==2.3.1
pyflakes==1.5.0
python-resize-image==1.1.11
PyYAML==3.12
requests==2.18.1
urllib3==1.21.1
